requirements.txt
openai

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Extras
alembic>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary
SQLAlchemy
alembic
